Be defensive with Principia
Principia doesn't guarantee that the values in the object passed from within
the realms of it's inconceivable depths are, or will remain to be, fields or
properties. So use the reflection method from their wiki, and care about
possible performance considerations later.

See also https://github.com/mockingbirdnest/Principia/pull/2566